refactor: split large files into focused single-responsibility modules
commands.rs (2478 lines) → src/cli/cmd/ (14 files):
  ui.rs       shared display helpers (is_tty, spinner, progress_style, etc.)
  index.rs    index command
  search.rs   search command + resolve_project_and_deps / search_all_dbs
  ask.rs      ask command + ask_json_schema
  status.rs   status command + format_age
  graph.rs    graph command + print_edges
  link.rs     link / unlink / autoclean commands
  misc.rs     languages / chunks commands
  check.rs    check command
  hooks.rs    hooks command + git-dir helpers
  memory.rs   memory command + all memory sub-handlers
  verify.rs   verify command
  plan.rs     plan command + plan_create / plan_status

parser.rs (739 lines) → src/indexer/parser/ (3 files):
  mod.rs        public API: SUPPORTED_LANGUAGES, detect_*, SourceParser
  ts_walker.rs  tree-sitter internals: NodeSpec, node_specs, walk_node,
                extract_name, all language-specific name extractors
  text.rs       parse_markdown, parse_notebook, atx_heading

graph.rs (669 lines) → src/indexer/graph/ (3 files):
  mod.rs       EdgeKind, EdgeExtractor, walk, enclosing_scope, collect
  edges.rs     per-language edge extraction (rust, python, js, go, java, c, html, css)
  builtins.rs  is_*_builtin filter sets

No behaviour changes — pure structural refactor. cargo build passes cleanly.
