+++ src/cli/cmd/check.rs
@@ -1,0 +1,66 @@
+use anyhow::Result;
+
+use crate::{
+    config::{Config, resolve_db},
+    storage::Database,
+};
+use super::super::CheckArgs;
+
+pub fn check(args: CheckArgs, cfg: Config) -> Result<()> {
+    let db_path = resolve_db(args.db.as_deref(), &cfg.db_path);
+    if !db_path.exists() {
+        anyhow::bail!(
+            "No index found (checked current directory and parents).\n\
+             Run `spelunk index <path>` first."
+        );
+    }
+
+    let db = Database::open(&db_path)?;
+    let stored = db.all_file_hashes()?;
+
+    let mut stale: Vec<String> = Vec::new();
+
+    // Check every indexed file against its current on-disk hash.
+    for (path, stored_hash) in &stored {
+        match std::fs::read(path) {
+            Ok(bytes) => {
+                let current = format!("{}", blake3::hash(&bytes));
+                if current != *stored_hash {
+                    stale.push(path.clone());
+                }
+            }
+            Err(_) => {
+                // File deleted since last index.
+                stale.push(path.clone());
+            }
+        }
+    }
+
+    let fmt = crate::utils::effective_format(&args.format);
+    let fresh = stale.is_empty();
+
+    if fmt == "json" {
+        println!(
+            "{}",
+            serde_json::to_string_pretty(&serde_json::json!({
+                "fresh": fresh,
+                "indexed_files": stored.len(),
+                "stale_files": stale.len(),
+                "stale": stale,
+            }))?
+        );
+    } else if fresh {
+        println!("Index is up to date. ({} files indexed)", stored.len());
+    } else {
+        println!("{} file(s) changed since last index:", stale.len());
+        for p in &stale {
+            println!("  {p}");
+        }
+        println!("\nRun `spelunk index .` to update.");
+    }
+
+    if !fresh {
+        std::process::exit(1);
+    }
+    Ok(())
+}
