+++ src/cli/cmd/mod.rs
@@ -1,0 +1,26 @@
+mod ui;
+mod index;
+mod search;
+mod ask;
+mod status;
+mod graph;
+mod link;
+mod misc;
+mod check;
+mod hooks;
+mod memory;
+mod verify;
+mod plan;
+
+pub use index::index;
+pub use search::search;
+pub use ask::ask;
+pub use status::status;
+pub use graph::graph;
+pub use link::{link, unlink, autoclean};
+pub use misc::{languages, chunks};
+pub use check::check;
+pub use hooks::hooks;
+pub use memory::memory;
+pub use verify::verify;
+pub use plan::plan;
