+++ src/cli/cmd/search.rs
@@ -1,0 +1,147 @@
+use anyhow::{Context, Result};
+
+use crate::{
+    config::{Config, resolve_db},
+    embeddings::{EmbeddingBackend as _, vec_to_blob},
+    registry::Registry,
+    search::SearchResult,
+    storage::Database,
+};
+use super::super::{SearchArgs};
+use super::ui::{spinner, print_results_text};
+
+pub async fn search(args: SearchArgs, cfg: Config) -> Result<()> {
+    let (db_path, dep_dbs) = resolve_project_and_deps(args.db.as_ref(), &cfg)?;
+
+    let sp = spinner("Loading model…");
+
+    let embedder = crate::backends::ActiveEmbedder::load(&cfg)
+        .await
+        .with_context(|| format!("loading embedding model '{}'", cfg.embedding_model))?;
+
+    sp.set_message("Embedding query…");
+    let query_text = format!("task: code retrieval | query: {}", args.query);
+    let vecs = embedder.embed(&[&query_text]).await?;
+    let query_blob = vec_to_blob(vecs.first().context("no embedding returned")?);
+
+    sp.set_message("Searching…");
+    let mut results = search_all_dbs(&db_path, &dep_dbs, &query_blob, args.limit.min(100))?;
+    sp.finish_and_clear();
+
+    if results.is_empty() {
+        println!("No results found. Make sure the index has embeddings (`spelunk index <path>`).");
+        return Ok(());
+    }
+
+    // ── Graph-aware enrichment (primary DB only) ──────────────────────────────
+    if args.graph
+        && let Ok(primary_db) = Database::open(&db_path)
+    {
+        let seen_ids: std::collections::HashSet<i64> = results.iter().map(|r| r.chunk_id).collect();
+        let names: Vec<&str> = results.iter().filter_map(|r| r.name.as_deref()).collect();
+
+        if !names.is_empty()
+            && let Ok(neighbor_ids) = primary_db.graph_neighbor_chunks(&names)
+        {
+            let new_ids: Vec<i64> = neighbor_ids
+                .into_iter()
+                .filter(|id| !seen_ids.contains(id))
+                .take(args.graph_limit)
+                .collect();
+
+            if !new_ids.is_empty()
+                && let Ok(mut extra) = primary_db.chunks_by_ids(&new_ids)
+            {
+                for r in &mut extra {
+                    r.from_graph = true;
+                }
+                results.extend(extra);
+            }
+        }
+    }
+
+    match crate::utils::effective_format(&args.format) {
+        "json" => println!("{}", serde_json::to_string_pretty(&results)?),
+        _ => print_results_text(&results),
+    }
+
+    Ok(())
+}
+
+/// Resolve the primary DB path and any dep DB paths via the registry.
+/// Falls back to `resolve_db` if the registry can't find a project.
+pub(crate) fn resolve_project_and_deps(
+    explicit_db: Option<&std::path::PathBuf>,
+    cfg: &Config,
+) -> Result<(std::path::PathBuf, Vec<std::path::PathBuf>)> {
+    // Explicit --db skips registry entirely.
+    if let Some(p) = explicit_db {
+        if !p.exists() {
+            anyhow::bail!(
+                "Database not found at '{}'. Run `spelunk index <path>` first.",
+                p.display()
+            );
+        }
+        return Ok((p.clone(), vec![]));
+    }
+
+    // Try registry first.
+    if let Ok(reg) = Registry::open()
+        && let Ok(cwd) = std::env::current_dir()
+        && let Ok(Some(project)) = reg.find_project_for_path(&cwd)
+        && project.db_path.exists()
+    {
+        let deps = reg
+            .get_deps(project.id)
+            .unwrap_or_default()
+            .into_iter()
+            .map(|d| d.db_path)
+            .filter(|p| p.exists())
+            .collect();
+        return Ok((project.db_path, deps));
+    }
+
+    // Fallback: filesystem walk-up.
+    let db_path = resolve_db(None, &cfg.db_path);
+    if !db_path.exists() {
+        anyhow::bail!(
+            "No index found (checked current directory and parents).\n\
+             Run `spelunk index <path>` inside your project first."
+        );
+    }
+    Ok((db_path, vec![]))
+}
+
+/// Search a primary DB and any dep DBs, merge results by distance, return top `limit`.
+pub(crate) fn search_all_dbs(
+    primary_db_path: &std::path::Path,
+    dep_db_paths: &[std::path::PathBuf],
+    query_blob: &[u8],
+    limit: usize,
+) -> Result<Vec<SearchResult>> {
+    let primary_db = Database::open(primary_db_path)?;
+    // Over-fetch from each DB so we have enough candidates after merging.
+    let fetch = (limit * 2).max(limit + 10);
+    let mut all = primary_db.search_similar(query_blob, fetch)?;
+
+    for dep_path in dep_db_paths {
+        match Database::open(dep_path) {
+            Ok(dep_db) => match dep_db.search_similar(query_blob, fetch) {
+                Ok(mut dep_results) => all.append(&mut dep_results),
+                Err(e) => tracing::warn!("search failed on dep {}: {e}", dep_path.display()),
+            },
+            Err(e) => tracing::warn!("could not open dep DB {}: {e}", dep_path.display()),
+        }
+    }
+
+    // Sort by distance (ascending), deduplicate by (file_path, start_line, end_line).
+    all.sort_by(|a, b| {
+        a.distance
+            .partial_cmp(&b.distance)
+            .unwrap_or(std::cmp::Ordering::Equal)
+    });
+    let mut seen = std::collections::HashSet::new();
+    all.retain(|r| seen.insert((r.file_path.clone(), r.start_line, r.end_line)));
+    all.truncate(limit);
+    Ok(all)
+}
