+++ src/main.rs
@@ -46,20 +46,20 @@ async fn main() -> Result<()> {
     let cfg = config::Config::load(cli.config.as_deref())?;
 
     match cli.command {
-        Command::Index(args) => cli::commands::index(args, cfg).await,
-        Command::Search(args) => cli::commands::search(args, cfg).await,
-        Command::Ask(args) => cli::commands::ask(args, cfg).await,
-        Command::Status(args) => cli::commands::status(args, cfg).await,
-        Command::Check(args) => cli::commands::check(args, cfg),
-        Command::Languages => cli::commands::languages(),
-        Command::Graph(args) => cli::commands::graph(args, cfg),
-        Command::Chunks(args) => cli::commands::chunks(args, cfg),
-        Command::Verify(args) => cli::commands::verify(args, cfg).await,
-        Command::Link(args) => cli::commands::link(args, cfg),
-        Command::Unlink(args) => cli::commands::unlink(args, cfg),
-        Command::Autoclean => cli::commands::autoclean(cfg),
-        Command::Memory(args) => cli::commands::memory(args, cfg).await,
-        Command::Hooks(args) => cli::commands::hooks(args),
-        Command::Plan(args) => cli::commands::plan(args, cfg).await,
+        Command::Index(args) => cli::cmd::index(args, cfg).await,
+        Command::Search(args) => cli::cmd::search(args, cfg).await,
+        Command::Ask(args) => cli::cmd::ask(args, cfg).await,
+        Command::Status(args) => cli::cmd::status(args, cfg).await,
+        Command::Check(args) => cli::cmd::check(args, cfg),
+        Command::Languages => cli::cmd::languages(),
+        Command::Graph(args) => cli::cmd::graph(args, cfg),
+        Command::Chunks(args) => cli::cmd::chunks(args, cfg),
+        Command::Verify(args) => cli::cmd::verify(args, cfg).await,
+        Command::Link(args) => cli::cmd::link(args, cfg),
+        Command::Unlink(args) => cli::cmd::unlink(args, cfg),
+        Command::Autoclean => cli::cmd::autoclean(cfg),
+        Command::Memory(args) => cli::cmd::memory(args, cfg).await,
+        Command::Hooks(args) => cli::cmd::hooks(args),
+        Command::Plan(args) => cli::cmd::plan(args, cfg).await,
     }
 }
